Adjust login page
